requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
openai==1.54.0
httpx==0.27.2
PyPDF2==3.0.1
python-docx==1.1.0
werkzeug==3.0.1
